+++ sparse/coo/__init__.py
@@ -1,4 +1,5 @@
 from .core import COO, as_coo
 from .umath import elemwise
-from .common import tensordot, dot, concatenate, stack, triu, tril, where, \
-    nansum, nanprod, nanmin, nanmax, nanreduce, roll
+from .common import (tensordot, dot, concatenate, stack, triu, tril, where,
+                     nansum, nanprod, nanmin, nanmax, nanreduce, roll, eye,
+                     zeros, zeros_like, ones, ones_like)